PPLJLS-8839 First take on the memory leak in IdentifierCacheHandler
The memory leak has been noticed in the asm-execution server. When the IdentifierCacheHandler.Key (that is a key of the IdentifierCacheHandler#idMap map) is an EObject, the idMap grows enormously because of the fact that we have no equals/hashCode methods overridden in our EMF objects. The result is that we have duplicated elements stored in this map which are not needed there - and probably are not being cleared by the garbage collector.

The solution here was to check if the key is an EObject and has the id attribute - and if that's the case, then compare the ids. If they are the same, that means we try to put the new version of the same object to the idMap (which makes sense since we do not need the older versions in the idMap map) - so thanks to the new implementation we put the new version of the object in the place of the older one.