PPLJLS-8839 Because of memory leaks the IdentifierCacheHandler class has been extended by supporting the EObject case
When the Key of the idMap is an EObject (that has no equals() method by design), each time a new version of this object is being added to this map, it is being added as a second entry (and it does not overwrite the old entry). This behavior has caused massive memory leaks (mostly in asm-execution) - so in order to fix that, IdentifierCacheHandler.Key#equals and IdentifierCacheHandler.Key#hashcode have been modified. Right now when the Key is an EObject and it has the 'id' attribute (which represents the database id), the uniqueness of the Key is being based on this id and the name of the given EObject. Thanks to it, we no longer have duplicates in the idMap.